prefer not to use "" for 0
(Should have probably posted this as an issue rather than a commit
like this, but hopefully this way is clear.)

I think using "" instead of 0 causes lots of unnecessary confusion,
and I haven't seen this before (maybe it is more common than I
know). For example, the first sentence after this table is "If _n_ is
even, then the least significant digit of _n_’s binary representation
is 0," but that's not actually true if 0 is represented as "", we need
a special case for it. (Maybe I'm missing some advantage of this
representation, but if not, seems like a lot of unnecessary complexity
rather than just using 0.)