string notation
I think its good to have the quotes around strings like this, but none
of the previous bitstrings used quotes. Another option to distinguish
strings from numbers if the quotes are cumbersome would be to use a
different (e.g., sans serif) font. This is why I had problems with the
earlier NtS definiiton, since juxtaposing a number font 0 with the
output of a function looks more like multiplication than string
concatenation.