using docstrings or not?
In the earlier Python examples, I assumed you wanted to avoid using
docstring comments, but here you use them for the inner functions but
not the main one. I'm not sure if there is a good reason for this, but
given that lots of the Python code is fairly Pythonic (e.g., using
Python-specific features like "".join ...), seems more natural to use
docstrings for all the examples.