minor edits
I didn't suggest edits for this, but find it a bit odd that the
example for Python programs is described as a partial function, but
the one where the input is a representation of a graph (where many
strings in {0, 1}* do not correspond to valid graphs) is not (even the
first example is also a partial function, since the pair
representation of the input is only valid for some strings). I'm not
sure this is such an important distinction to make here, but if it is,
should make it for the other.

The second one can be total with the restriction I added (input must
be greater than 1).