Multiple small changes
----preface.md ----
"for" seems more clear to me than "in" (1st change)

"how much better you can understand a proof" seems somewhat less awkward to me than "how much you can understand a proof better"

"what the natural examples are of objects" seems less awkward to me than "what are natural examples of objects", which sounds more like a question

The @Shannon1937 bibliographical reference did not render properly on the html. This appears to be due to the entry in the tcs.bib file being @Shannon1938, not @Shannon1937. I looked up the paper and 1938 is the correct year.

I replaced "software defined" with "software-defined" as it seems less awkward, and this is consistent with the phrasing used in the referenced paper

Changed two "two-dot" ellipses (..) to the generally accepted "three-dot" ellipses (...). I could not find precedent for a two-dot ellipsis, I apologize if this is simply author's preference.

README.md
Fixed a double space. Not that that really matters at all, since it's just a readme.