Removed footnotes from blockquotes at beginning of lec_01
Looking at the html-rendered version of lec_01, when clicking on the "..." footnote buttons, the spawned footnote textboxes have a transparent background leading them to overlap over main text in a disorienting way.

When the markdown is rendered, the footnote textbox (once spawned) is contained within a <blockquote> element - I believe that the bug is a result of the footnote inheriting some sort of property from the blockquote element it is within, but it doesn't seem to be a simply `opacity` style, as modifying those in chrome dev tools did not appear to change the appearance of the footnote text box.

I spent a couple hours trying to make the markdown rendering scripts to work to test things, but unfortunately I was not successful. Simply moving the DOM element of the footnote spawning button outside of the blockquote in the rendered html did not appear to fix things - the transparency was gone but this led to other severe visual bugs.

I noticed in your README that you noted that one of your priorities was to "Make the text more "linear": fewer footnotes and remarks that disrupt the flow of reading."

Given that one of your goals is removing footnotes, this visual bug makes the footnote very difficult to read, and it appears a true fix might be difficult, the simplest solution seems to be to simply remove them.