Grammatical fix: without the commas around "or recipe" on line 58, the word recipe would require an article ("or a recipe"), but that would imply that an algorithm is either a set of instructions OR a recipe, when really they're equivalent.
Changing the phrase "a set of instructions or recipe" to "a set of instructions, or recipe, " maintains what I believe was the author's intent of simply using a colloquial term ("recipe") to allow the student to more intuitively "set of instructions", while being grammatically correct.

The grammatical error was confirmed with grammarly Pro, the best grammar checker I could think of.