In the current version, the naive multiplication algorithm and the "grade school multiplication" algorithm are defined immediately after a sentence "We start by formally describing both algorithms, see Algorithm 0.1 and Algorithm 0.2."
Algorithm 0.1 and Algorithm 0.2 are relative hyperlinks (to "#naivemultalg"  and "#gradeschoolalg")

There are two problems with this: first, these links don't lead anywhere. While the markdown source file does contain the relative link target in the algorithm definition ({.algorithm title="Multiplication via repeated addition" #naivemultalg}, for example), this information seems to somehow be lost in the rendering. Opening the relative links in a new tab simply puts you at the top of the page, as if you'd loaded it without the link.

Second, the link text is "Algorithm 0.x" while the text in the latex heading is "Algorithm x" . While I understand that the 0. prefix simply identifies what chapter we're in, (and I genuinely appreciate that, as many books are pretty care-free about referencing the chapter), consistency between the two in the chapter could be preferable.